Re-write of serialisation code: converting directly between JSON and BSON, without use of Go structs for all elements.
Primitive extensions now supported (JSON only).
Ranges stored for dates & decimals and used for searches.
Decimals stored with original representation.
(Not backwards compatible with previous data)

Temporary broken search for quantity with '||' syntax (on both code and unit)

Test fixtures updated to STU3

+++ models/quantity.go
@@ -27,7 +27,7 @@
 package models
 
 type Quantity struct {
-	Value      *float64 `bson:"value,omitempty" json:"value,omitempty"`
+	Value      *Decimal `bson:"value,omitempty" json:"value,omitempty"`
 	Comparator string   `bson:"comparator,omitempty" json:"comparator,omitempty"`
 	Unit       string   `bson:"unit,omitempty" json:"unit,omitempty"`
 	System     string   `bson:"system,omitempty" json:"system,omitempty"`
